refactor: 누적 리워드 합계 필드를 long으로 수정 (#175)

+++ src/main/java/server/MATE/domain/answer/dto/response/MyAnswerResponse.java
@@ -3,10 +3,10 @@ package server.MATE.domain.answer.dto.response;
 import java.util.List;
 
 public record MyAnswerResponse(
-        Integer totalPromotionReward,
+        Long totalPromotionReward,
         List<MyAnswerItem> answers
 ) {
-    public static MyAnswerResponse of(Integer totalPromotionReward, List<MyAnswerItem> answers) {
+    public static MyAnswerResponse of(Long totalPromotionReward, List<MyAnswerItem> answers) {
         return new MyAnswerResponse(totalPromotionReward, answers);
     }
 }
